status 型 integer へ変更

+++ lib/models/swagger/v1/model/request_task_put_schema.dart
@@ -5,7 +5,7 @@ class RequestTaskPutSchema {
   
   String? name;
 
-  double? status;
+  int? status;
 
   RequestTaskPutSchema();
